[fontc] Replace test_args with Args::for_test
A little more cleanup.
